refactor(eval): extract runner config helpers
