+++ src/eval/runner.ts
@@ -1,22 +1,20 @@
 import { existsSync } from "fs";
-import { mkdirSync } from "fs";
-import { readFileSync } from "fs";
-import { rmSync } from "fs";
-import { writeFileSync } from "fs";
-import * as os from "os";
 import * as path from "path";
 import { performance } from "perf_hooks";
 
-import { rebasePathEntries, resolveInheritedKnowledgeBaseEntries } from "../config/rebase.js";
-import { parseConfig } from "../config/schema.js";
-import type { SearchConfig as ConfigSearchConfig } from "../config/schema.js";
-import { getDefaultModelForProvider } from "../config/index.js";
-import { getGlobalIndexPath, resolveProjectConfigPath, resolveProjectIndexPath } from "../config/paths.js";
 import { Indexer } from "../indexer/index.js";
 
 import { evaluateBudgetGate } from "./budget.js";
 import { compareSummaries } from "./compare.js";
 import { buildPerQueryResult, computeEvalMetrics } from "./metrics.js";
+import {
+  clearIndexRoot,
+  ensureLocalEvalProjectConfig,
+  getEmbeddingCostPer1MTokens,
+  loadParsedConfig,
+  resolveSearchConfig,
+  toAbsolute,
+} from "./runner-config.js";
 import {
   createSummaryMarkdown,
   createRunDirectory,
@@ -36,150 +34,6 @@ import type {
   SweepDefinition,
   SweepRunSummary,
 } from "./types.js";
-
-function toAbsolute(projectRoot: string, maybeRelative: string): string {
-  return path.isAbsolute(maybeRelative) ? maybeRelative : path.join(projectRoot, maybeRelative);
-}
-
-function isProjectScopedConfigPath(configPath: string): boolean {
-  return path.basename(configPath) === "codebase-index.json"
-    && path.basename(path.dirname(configPath)) === ".opencode";
-}
-
-function normalizeEvalConfigKnowledgeBases(
-  rawConfig: unknown,
-  projectRoot: string,
-  resolvedConfigPath: string,
-): Record<string, unknown> {
-  const config = rawConfig && typeof rawConfig === "object"
-    ? { ...(rawConfig as Record<string, unknown>) }
-    : {};
-
-  if (!Array.isArray(config.knowledgeBases)) {
-    return config;
-  }
-
-  config.knowledgeBases = isProjectScopedConfigPath(resolvedConfigPath)
-    ? resolveInheritedKnowledgeBaseEntries(
-        config.knowledgeBases,
-        path.dirname(path.dirname(resolvedConfigPath)),
-        projectRoot,
-      )
-    : rebasePathEntries(
-        config.knowledgeBases,
-        path.dirname(resolvedConfigPath),
-        projectRoot,
-      );
-
-  return config;
-}
-
-function loadRawConfig(projectRoot: string, configPath?: string): unknown {
-  const fromPath = configPath ? toAbsolute(projectRoot, configPath) : null;
-  if (fromPath && existsSync(fromPath)) {
-    return normalizeEvalConfigKnowledgeBases(
-      JSON.parse(readFileSync(fromPath, "utf-8")),
-      projectRoot,
-      fromPath,
-    );
-  }
-
-  const projectConfig = resolveProjectConfigPath(projectRoot);
-  if (existsSync(projectConfig)) {
-    return normalizeEvalConfigKnowledgeBases(
-      JSON.parse(readFileSync(projectConfig, "utf-8")),
-      projectRoot,
-      projectConfig,
-    );
-  }
-
-  const globalConfig = path.join(os.homedir(), ".config", "opencode", "codebase-index.json");
-  if (existsSync(globalConfig)) {
-    return JSON.parse(readFileSync(globalConfig, "utf-8"));
-  }
-
-  return {};
-}
-
-function getIndexRootPath(projectRoot: string, scope: "project" | "global"): string {
-  return scope === "global"
-    ? getGlobalIndexPath()
-    : resolveProjectIndexPath(projectRoot, scope);
-}
-
-function getLocalProjectIndexRoot(projectRoot: string): string {
-  return path.join(projectRoot, ".opencode", "index");
-}
-
-function getLocalProjectConfigPath(projectRoot: string): string {
-  return path.join(projectRoot, ".opencode", "codebase-index.json");
-}
-
-function clearIndexRoot(projectRoot: string, scope: "project" | "global"): void {
-  const indexRoot = scope === "global"
-    ? getIndexRootPath(projectRoot, scope)
-    : getLocalProjectIndexRoot(projectRoot);
-  if (existsSync(indexRoot)) {
-    rmSync(indexRoot, { recursive: true, force: true });
-  }
-}
-
-function ensureLocalEvalProjectConfig(projectRoot: string, configPath?: string): string | undefined {
-  const localConfigPath = getLocalProjectConfigPath(projectRoot);
-  const resolvedConfigPath = configPath
-    ? toAbsolute(projectRoot, configPath)
-    : resolveProjectConfigPath(projectRoot);
-
-  if (!configPath && existsSync(localConfigPath)) {
-    return localConfigPath;
-  }
-
-  if (!existsSync(resolvedConfigPath) || resolvedConfigPath === localConfigPath) {
-    return resolvedConfigPath;
-  }
-
-  const sourceConfig = normalizeEvalConfigKnowledgeBases(
-    JSON.parse(readFileSync(resolvedConfigPath, "utf-8")),
-    projectRoot,
-    resolvedConfigPath,
-  );
-
-  mkdirSync(path.dirname(localConfigPath), { recursive: true });
-  writeFileSync(localConfigPath, JSON.stringify(sourceConfig, null, 2), "utf-8");
-  return localConfigPath;
-}
-
-function loadParsedConfig(projectRoot: string, configPath?: string) {
-  const raw = loadRawConfig(projectRoot, configPath);
-  return parseConfig(raw);
-}
-
-function resolveSearchConfig(
-  parsedConfig: ReturnType<typeof parseConfig>,
-  overrides?: Partial<Pick<ConfigSearchConfig, "fusionStrategy" | "hybridWeight" | "rrfK" | "rerankTopN">>
-) {
-  const nextSearch: ConfigSearchConfig = {
-    ...parsedConfig.search,
-  };
-
-  if (overrides?.fusionStrategy !== undefined) {
-    nextSearch.fusionStrategy = overrides.fusionStrategy;
-  }
-  if (overrides?.hybridWeight !== undefined) {
-    nextSearch.hybridWeight = overrides.hybridWeight;
-  }
-  if (overrides?.rrfK !== undefined) {
-    nextSearch.rrfK = overrides.rrfK;
-  }
-  if (overrides?.rerankTopN !== undefined) {
-    nextSearch.rerankTopN = overrides.rerankTopN;
-  }
-
-  return {
-    ...parsedConfig,
-    search: nextSearch,
-  };
-}
 
 export interface EvalRunResult {
   outputDir: string;
@@ -243,10 +97,7 @@ export async function runEvaluation(options: EvalRunOptions): Promise<EvalRunRes
   const logger = indexer.getLogger();
   const metricSnapshot = logger.getMetrics();
 
-  const costPer1MTokensUsd =
-    effectiveConfig.embeddingProvider === "custom" || effectiveConfig.embeddingProvider === "auto"
-      ? 0
-      : getDefaultModelForProvider(effectiveConfig.embeddingProvider).costPer1MTokens;
+  const costPer1MTokensUsd = getEmbeddingCostPer1MTokens(effectiveConfig.embeddingProvider);
 
   const summary: EvalSummary = {
     generatedAt: new Date().toISOString(),
